#57 misc tidy
