+++ src/universal/conf/application.ini
@@ -1,5 +1,5 @@
 
-# This script will be pulled into the main script via the folllwing 
+# This script will be pulled into the main script via the following 
 # (found in loadConfigFile). 
 #
 #   cat application.ini | grep sed '/^\#/d;s/\r$//'
